move some more files and fix tests
