+++ packages/live-share/src/index.ts
@@ -21,18 +21,15 @@ export * from "./TestLiveShareHost";
 export * from "./TimeInterval";
 export * from "./TimestampProvider";
 
-// TODO: our package structure is such that things that need to be exposed at all for other live share libs
-// are in the root of the directory even if they are "internal". move some to internal again?
+// Exposed for other live share libs, or advanced use cases.
 export * from "./internals/BaseLiveShareClient";
 export * from "./internals/DynamicObjectRegistry";
 export * from "./internals/DynamicObjectManager";
+export * from "./internals/LiveDataObject";
+export * from "./internals/LiveEventScope";
+export * from "./internals/LiveEventSource";
+export * from "./internals/LiveEventTarget";
+export * from "./internals/LiveObjectSynchronizer";
+export * from "./internals/LiveShareRuntime";
 export * from "./internals/schema-injection-utils";
 export * from "./internals/smuggle";
-
-// TODO: move to internal
-export * from "./LiveDataObject";
-export * from "./LiveEventScope";
-export * from "./LiveEventSource";
-export * from "./LiveEventTarget";
-export * from "./LiveObjectSynchronizer";
-export * from "./LiveShareRuntime";
