+++ packages/live-share/src/internals/DynamicObjectManager.ts
@@ -16,7 +16,7 @@ import { IFluidHandle, IFluidLoadable } from "@fluidframework/core-interfaces";
 import { assert } from "@fluidframework/core-utils/internal";
 import { ConsensusRegisterCollection } from "@fluidframework/register-collection/internal";
 import { DynamicObjectRegistry } from "./DynamicObjectRegistry";
-import { LiveDataObject } from "../LiveDataObject";
+import { LiveDataObject } from "./LiveDataObject";
 
 // Register ConsensusRegisterCollection
 DynamicObjectRegistry.registerObjectClass(
