+++ packages/live-share/src/internals/LiveEventScope.ts
@@ -9,12 +9,12 @@ import {
 } from "@fluidframework/core-interfaces";
 import { TypedEventEmitter } from "@fluid-internal/client-utils";
 import { IInboundSignalMessage } from "@fluidframework/runtime-definitions/internal";
-import { ILiveEvent, UserMeetingRole } from "./interfaces";
+import { ILiveEvent, UserMeetingRole } from "../interfaces";
 import { LiveShareRuntime } from "./LiveShareRuntime";
 import { IEvent } from "@fluidframework/core-interfaces";
-import { LiveTelemetryLogger } from "./LiveTelemetryLogger";
-import { isILiveEvent } from "./internals/type-guards";
-import { waitUntilConnected } from "./internals/utils";
+import { LiveTelemetryLogger } from "../LiveTelemetryLogger";
+import { isILiveEvent } from "./type-guards";
+import { waitUntilConnected } from "./utils";
 
 /**
  * Live event callback.
