+++ packages/live-share/src/internals/LiveEventTarget.ts
