+++ packages/live-share/src/internals/LiveObjectSynchronizer.ts
@@ -9,7 +9,7 @@ import {
     GetLocalUserCanSend,
     ILiveEvent,
     UpdateSynchronizationState,
-} from "./interfaces";
+} from "../interfaces";
 
 /**
  * @internal
