+++ packages/live-share/src/internals/test/MockRuntimeSignaler.ts
@@ -3,7 +3,7 @@
  * Licensed under the Microsoft Live Share SDK License.
  */
 
-import { IInboundSignalMessage } from "@fluidframework/runtime-definitions";
+import { IInboundSignalMessage } from "@fluidframework/runtime-definitions/internal";
 import { MockLogger } from "@fluidframework/telemetry-utils/internal";
 import { IRuntimeSignaler } from "../LiveEventScope";
 import { v4 } from "uuid";
