+++ packages/live-share/src/test/LiveEvent.spec.ts
@@ -10,15 +10,15 @@ import {
     getContainerEntryPointBackCompat,
 } from "@fluidframework/test-utils/internal";
 import { LiveEvent } from "../LiveEvent";
-import { Deferred } from "../internals";
 import { MockTimestampProvider } from "./MockTimestampProvider";
 import { MockRoleVerifier } from "./MockRoleVerifier";
 import { LocalTimestampProvider } from "../LocalTimestampProvider";
 import { UserMeetingRole, ILiveEvent } from "../interfaces";
 import { TestLiveShareHost } from "../TestLiveShareHost";
 import { getLiveDataObjectKind } from "../internals/schema-injection-utils";
-import { LiveShareRuntime } from "../LiveShareRuntime";
+import { LiveShareRuntime } from "../internals/LiveShareRuntime";
 import { describeCompat } from "@live-share-private/test-utils";
+import { Deferred } from "../internals/Deferred";
 
 describeCompat("LiveEvent", (getTestObjectProvider) => {
     let provider: ITestObjectProvider;
